Segmenting and Clustering Neighborhoods
exploring neighborhoods in New York City to get the most common venue categories in each neighborhood and then use this feature to group the neighborhoods into clusters using the k-means clustering algorithm. Foursquare API is used to explore neighborhoods in New York City and Folium library is used to visualize the neighborhoods in New York City and their emerging clusters.